LPS-63232 Move portlet keys into API
