+++ modules/apps/invitation/invitation-invite-members-api/src/main/java/com/liferay/invitation/invite/members/constants/InviteMembersPortletKeys.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.invitation.invite.members.web.constants;
+package com.liferay.invitation.invite.members.constants;
 
 /**
  * @author Adolfo Pérez
