+++ modules/apps/invitation/invitation-invite-members-web/src/main/java/com/liferay/invitation/invite/members/web/notifications/InviteMembersUserNotificationHandler.java
@@ -14,9 +14,9 @@
 
 package com.liferay.invitation.invite.members.web.notifications;
 
+import com.liferay.invitation.invite.members.constants.InviteMembersPortletKeys;
 import com.liferay.invitation.invite.members.model.MemberRequest;
 import com.liferay.invitation.invite.members.service.MemberRequestLocalService;
-import com.liferay.invitation.invite.members.web.constants.InviteMembersPortletKeys;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.model.Group;
